[Edit] 로그인 서비스 통합 (#129)
